cargo clippy --fix --workspace --all-features

+++ pallets/subtensor/src/tests/coinbase.rs
@@ -33,10 +33,10 @@ fn test_dynamic_function_various_values() {
                     SubnetTAO::<Test>::insert(1, (price * 1_000_000_000.0) as u64);
                     SubnetAlphaIn::<Test>::insert(1, 1_000_000_000);
                     let (tao_in_emission, alpha_in_emission, alpha_out_emission) = SubtensorModule::get_dynamic_tao_emission( 1, tao_in, alpha_emission);
-                    assert!(tao_in_emission <= tao_in as u64, "tao_in_emission is greater than tao_in");
-                    assert!(alpha_in_emission <= alpha_emission as u64, "alpha_in_emission is greater than alpha_emission");
-                    assert!(alpha_out_emission <= 2 * alpha_emission as u64, "alpha_out_emission is greater than 2 * alpha_emission");
-                    assert!((alpha_in_emission + alpha_out_emission) <= 2 * alpha_emission as u64, "Sum of alpha_in_emission and alpha_out_emission is less than or equal to. 2 * alpha_emission");
+                    assert!(tao_in_emission <= tao_in, "tao_in_emission is greater than tao_in");
+                    assert!(alpha_in_emission <= alpha_emission, "alpha_in_emission is greater than alpha_emission");
+                    assert!(alpha_out_emission <= 2 * alpha_emission, "alpha_out_emission is greater than 2 * alpha_emission");
+                    assert!((alpha_in_emission + alpha_out_emission) <= 2 * alpha_emission, "Sum of alpha_in_emission and alpha_out_emission is less than or equal to. 2 * alpha_emission");
                     close( alpha_in_emission + alpha_out_emission, 2 * alpha_emission, 10 );
                     if alpha_in_emission > 0 || tao_in_emission > 0 {
                         assert!((tao_in_emission as f64 / alpha_in_emission as f64 - price).abs() < 1e-1, "Ratio of tao_in_emission to alpha_in_emission is not equal to price");
